Improves the Login system
Login is still done with username and password, however employee_id is
returned, along with the access_token, which is used to identify the person
who is currently logged in. The Retrofit stuff as well as the existing
stub navigation functions have been changed to reflect the updated
system.

The access_token is passed as a header in every API call, apart from
those to /auth.